feat: make zoom configurable
